Factor out webRequest utility function
